aidsec_dashboard/requirements.txt
# AidSec Lead Dashboard Dependencies
streamlit>=1.33.0
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
plotly>=5.18.0

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
slowapi>=0.1.9
